Add Metro configuration for native and web apps
Set up custom Metro bundler configs for native and web apps in the monorepo. Adjusted dependencies in `package.json` to align with the new structure and support shared packages. Updated `yarn.lock` to reflect the changes.  This enables hot reload for both the native and web applications for the monorepo pacakges